+++ platform/swift/source/Logger.swift
@@ -88,24 +88,6 @@ public final class Logger {
         )
     }
 
-    convenience init?(
-        withAPIKey apiKey: String,
-        configuration: Configuration,
-        sessionConfiguration: SessionConfiguration = .init(),
-        dateProvider: DateProvider?,
-        fieldProviders: [FieldProvider],
-        loggerBridgingFactoryProvider: LoggerBridgingFactoryProvider = LoggerBridgingFactory()
-    ) {
-        self.init(
-            withAPIKey: apiKey,
-            configuration: configuration,
-            sessionStrategy: .configuration(sessionConfiguration),
-            dateProvider: dateProvider,
-            fieldProviders: fieldProviders,
-            loggerBridgingFactoryProvider: loggerBridgingFactoryProvider
-        )
-    }
-
     // swiftlint:disable function_body_length
     /// Internal constructor shared between the public convenience initializers and tests. Generally
     /// production apps would not pass in a bufferDirectory, it would default to the Capture default.
@@ -439,7 +421,7 @@ extension Logger: Logging {
 
     /// Creates a new session with an SDK-created ID.
     public func startNewSession() {
-        self.underlyingLogger.startNewSession(sessionID: nil)
+        self.startNewSession(sessionID: nil)
     }
 
     /// Creates a new session with an optional app-provided ID.